+++ CHANGELOG.md
@@ -30,6 +30,8 @@ resolve and bundle transitive dependencies of the upgraded `react-select` librar
 * Cube `View` now detects and properly handles streaming updates to source data that include changes
   to row dimensions as well as measures.*
 * `DataViewModel.itemHeight` can now be a function that returns a pixel height.
+* New `autoHeight` property on grid `Column`.  When set the grid will increase the row
+  height dynamically to accommodate cell content in this column.
 
 ### 📚 Libraries